changed to use u/v associated lon/lat instead of elevation's in tpxo
